로그인 그룹 배정 기능 완성, 회원탈퇴 기능 개선(1차)

+++ app/src/main/java/com/example/searchyourstuffeasily/ui/notifications/RevokeFragment.java
@@ -18,9 +18,10 @@ import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 
 public class RevokeFragment extends DialogFragment {
-    public FirebaseAuth mAuth;
-    DatabaseReference mDatabase= FirebaseDatabase.getInstance().getReference(); // 22
-    DatabaseReference userID = mDatabase.child("UserDB");
+    String famId = LoginActivity.usersFamilyId;
+    DatabaseReference HomeDbRef = FirebaseDatabase.getInstance().getReference("HomeDB");
+    DatabaseReference homesRef = FirebaseDatabase.getInstance().getReference("homes");
+    DatabaseReference usersRef = FirebaseDatabase.getInstance().getReference("users");
 
     @Override
     public Dialog onCreateDialog(Bundle savedInstanceState) {
@@ -28,14 +29,15 @@ public class RevokeFragment extends DialogFragment {
         builder.setMessage("회원탈퇴 하시겠습니까?")
                 .setPositiveButton("네", new DialogInterface.OnClickListener() {
                     public void onClick(DialogInterface dialog, int id) {
+                        //사용자의 familyId와 연결된 내용 삭제 후 로그아웃 처리
+                        deleteUser();
+                        FirebaseAuth.getInstance().signOut();
                         Intent intent = new Intent(getActivity().getApplicationContext(), LoginActivity.class);
                         startActivity(intent);
-                        deleteUser();
                     }
                 })
                 .setNegativeButton("아니오", new DialogInterface.OnClickListener() {
                     public void onClick(DialogInterface dialog, int id) {
-                        // User cancelled the dialog -> 기능 구현이 필요함(일자:24/05/10)
                     }
                 });
         return builder.create();
@@ -43,19 +45,18 @@ public class RevokeFragment extends DialogFragment {
 
     public void deleteUser() {
         FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
-        user.delete()
-                .addOnCompleteListener(new OnCompleteListener<Void>() {
-                    @Override
-                    public void onComplete(@NonNull Task<Void> task) {
-                        if (task.isSuccessful()) {
-                            //병합 후 오류 처리를 위해 임시 비활성화(일자:24/04/18)
-                            //활성화 시 LoginActivity에 context_login이 없어 오류가 발생하니 브랜치 병합 과정 중 다시 확인 필요(일자:24/05/10)
-                            //String cuid = ((LoginActivity)LoginActivity.context_login).cu;
-                            //userID.child(cuid).removeValue();
+        if(user != null){
+            user.delete()
+                    .addOnCompleteListener(new OnCompleteListener<Void>() {
+                        @Override
+                        public void onComplete(@NonNull Task<Void> task) {
+                            if (task.isSuccessful()) {
+                                HomeDbRef.child(famId).removeValue();
+                                homesRef.child(famId).removeValue();
+                                usersRef.child(user.getUid()).removeValue();
+                            }
                         }
-                    }
-                });
+                    });
+        }
     }
 }
-
-
